+++ eslint-local-rules.js
@@ -1,0 +1,3 @@
+module.exports = {
+  'forbid-use-hook-producer': require('./eslint-rules/forbid-use-hook-producer')
+};