+++ packages/component/src/Attachment/Text/private/CitationModalContent.tsx
@@ -1,4 +1,4 @@
-import { validateProps } from 'botframework-webchat-api/internal';
+import { validateProps } from 'botframework-webchat-react-valibot';
 import classNames from 'classnames';
 import React, { Fragment, memo, useMemo } from 'react';
 import { object, optional, pipe, readonly, string, type InferInput } from 'valibot';