+++ packages/component/src/index.ts
@@ -65,6 +65,14 @@ const hooks = {
   ...componentHooks
 };
 
+declare const process: {
+  env: {
+    build_tool: string;
+    module_format: string;
+    npm_package_version: string;
+  };
+};
+
 const buildTool = process.env.build_tool;
 const moduleFormat = process.env.module_format;
 const version = process.env.npm_package_version;