+++ packages/component/src/providers/GroupedRenderingActivities/GroupedRenderingActivitiesComposer.tsx
@@ -1,9 +1,9 @@
 import { hooks } from 'botframework-webchat-api';
 import { type WebChatActivity } from 'botframework-webchat-core';
+import { reactNode } from 'botframework-webchat-react-valibot';
 import React, { memo, useMemo } from 'react';
 import { object, optional, parse, pipe, readonly, type InferOutput } from 'valibot';
 
-import reactNode from '../../types/internal/reactNode';
 import useRenderingActivities from '../RenderingActivities/useRenderingActivities';
 import { type GroupedRenderingActivities } from './GroupedRenderingActivities';
 import GroupedRenderingActivitiesContext, {