+++ packages/react-valibot/src/reactNode.ts
@@ -10,7 +10,11 @@ function reactNode<
 function reactNode<
   const TMessage extends ErrorMessage<CustomIssue> | undefined = ErrorMessage<CustomIssue> | undefined
 >(message?: TMessage): CustomSchema<ReactNode, TMessage> {
-  return custom<ReactNode, TMessage>(() => true, message);
+  return custom<ReactNode, TMessage>(
+    () => true,
+    // TODO: Probably lacking some undefined checks, thus, we need to force cast.
+    message as TMessage
+  );
 }
 
 export default reactNode;