+++ packages/redux-store/src/index.ts
@@ -1,0 +1,2 @@
+export { default as ReduxStoreComposer } from './ReduxStoreComposer';
+export { default as useSuggestedActionsHooks } from './suggestedActions/useSuggestedActionsHooks';