+++ packages/redux-store/src/suggestedActions/useSuggestedActionsHooks.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import SuggestedActionsContext, { type SuggestedActionsContextType } from './private/SuggestedActionsContext';
+
+export default function useSuggestedActionsHooks(): SuggestedActionsContextType {
+  return useContext(SuggestedActionsContext);
+}